[feature] load and save vectors/matrices in textfiles (.txt or .mat)
This uses the load() and save('-acii',...) functions from octave.
Additionally, the behavior is slightly changed regarding the decision to
print to screen or save to file: before, if the output was a scalar, is
was printed to stdout. To allow for saving scalar results in text files,
this is now restricted to the cases where output=='-', that is, when the
option '-o' is not specified.
Also cleaned the prog variable to remove trailing whitespaces and
semicolons.